Success!  vewpoints now has brushes.  They require further enhancement and testing, but the hard part of tearing the code apart and putting it back togther is done.
Nearly every file was modified a little bit.  This is either the mark
of a major increase in functionality, or the mark of a poorly
organized code.  In this case, I think its a little of each.

All the brushes currently start out painting with "small, dim, red
square points", but they can be individually changed.  

New controls can be added "easily" for more brush features, and there
will be a broadcast group for making changes to all brushes at
once... etc.  The first enhancement I'll make is to have each brush
(there are eight) start off as a different color.



git-svn-id: http://svn2.assembla.com/svn/viewpoints@166 cd75e15f-47f5-0310-b60f-c4e69339362a